feat: extractor for customizations
Added extraction for property setters and links
